refactor: refactor folder structure of reusable components
