+++ src/common/components/BaseModal.vue
@@ -43,7 +43,7 @@
 </template>
 
 <script>
-import BaseButton from '@/components/ui/BaseButton.vue';
+import BaseButton from '@/common/components/BaseButton';
 
 export default {
   name: 'BaseModal',
