+++ src/common/components/SearchBar.vue
