+++ tests/unit/components/ui/Sidebar.spec.js
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils';
 import { NAVIGATION_MENU } from '@/static/data';
-import Sidebar from '@/components/ui/Sidebar';
+import Sidebar from '@/common/components/Sidebar';
 
 describe('Component: Sidebar', () => {
   let wrapper;
